Create PostCardsList component and lift state to HomePage component
